+++ gulp/flags.js
@@ -20,7 +20,8 @@ module.exports = function(options) {
             OUTBOUND_PAGES: argv.outboundPages || true,
             SHOW_ZENDESK: argv.showZendesk || true,
             APPDOCK: argv.showAppdock || true,
-            API_KEYS: argv.showApiKeys || false
+            API_KEYS: argv.showApiKeys || false,
+            CONTACT_MANAGEMENT: argv.contactManagement || true
           }
         }
       }))